refactors processPlugins, removes pluginReducers appendage from store
